python_cpp_opencv互调用/NtyVideoFullAutomatic10_12/requirements.txt
absl-py==0.10.0
altgraph==0.17
astunparse==1.6.3
baidu-aip==2.2.18.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cycler==0.10.0
future==0.18.2
gast==0.3.3
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
joblib==0.14.1
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.2.2
matplotlib==3.0.3
modbus-tk==1.1.0
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.3.0
pefile==2019.4.18
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyInstaller==4.0
pyparsing==2.4.7
PyQt5==5.13.0
PyQt5-sip==12.7.1
pyqt5-tools==5.13.0.1.5
pyserial==3.4
pytesseract==0.3.5
python-dateutil==2.8.1
pywin32==228
pywin32-ctypes==0.2.0
QCandyUi==0.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.14.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0
zope.event==4.4
zope.interface==5.1.0
